Hola

+++ BESA/AgenteResPwA/ResPwAEntities/src/ResPwAEntities/Joint.java
@@ -50,7 +50,7 @@ public class Joint implements Serializable {
     private double angulo;
     @Basic(optional = false)
     @Column(name = "TIEMPO")
-    private BigInteger tiempo;
+    private BigDecimal tiempo;
     @ManyToMany(mappedBy = "jointList", fetch = FetchType.EAGER)
     private List<Accion> accionList;
 
@@ -61,7 +61,7 @@ public class Joint implements Serializable {
         this.id = id;
     }
 
-    public Joint(BigDecimal id, String nombre, double angulo, BigInteger tiempo) {
+    public Joint(BigDecimal id, String nombre, double angulo, BigDecimal tiempo) {
         this.id = id;
         this.nombre = nombre;
         this.angulo = angulo;
@@ -92,11 +92,11 @@ public class Joint implements Serializable {
         this.angulo = angulo;
     }
 
-    public BigInteger getTiempo() {
+    public BigDecimal getTiempo() {
         return tiempo;
     }
 
-    public void setTiempo(BigInteger tiempo) {
+    public void setTiempo(BigDecimal tiempo) {
         this.tiempo = tiempo;
     }
